src/test/java/ru/javaops/topjava2/web/menu/UserMenuControllerTest.java
package ru.javaops.topjava2.web.menu;

import org.junit.jupiter.api.Test;
import org.springframework.http.MediaType;
import org.springframework.security.test.context.support.WithUserDetails;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import ru.javaops.topjava2.web.AbstractControllerTest;

import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static ru.javaops.topjava2.web.menu.MenuTestData.MENU_MATCHER;
import static ru.javaops.topjava2.web.menu.MenuTestData.menu2;
import static ru.javaops.topjava2.web.user.UserTestData.USER_MAIL;

class UserMenuControllerTest extends AbstractControllerTest {

    @Test
    @WithUserDetails(value = USER_MAIL)
    void getTodayByRestaurantId() throws Exception {
        perform(MockMvcRequestBuilders.get(UserMenuController.REST_URL + "/restaurant/2"))
                .andExpect(status().isOk())
                .andDo(print())
                .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
                .andExpect(MENU_MATCHER.contentJson(menu2));
    }
}